Desktop/优化后的niubiAI/requirements.py312.txt
# Core
python-telegram-bot==21.7
pydantic==2.10.6
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.44
asyncpg==0.29.0
alembic==1.13.1

# HTTP Client
httpx==0.28.1

# LLM APIs
openai==1.106.1
anthropic==0.40.0
google-generativeai==0.3.2

# Utils
tenacity==9.0.0
structlog==25.4.0

# Development
pytest==8.4.2
pytest-asyncio==0.23.8
black==25.1.0
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Optional dependencies (currently unused)
# Uncomment if needed:
# matplotlib==3.9.4
# numpy==2.0.2
# pandas==2.3.2
# Pillow==11.3.0
# yfinance==0.2.65
# tushare==1.4.24